+++ src/main/java/org/spring/attraction/service/AttractionTypeService.java
@@ -1,9 +1,14 @@
 package org.spring.attraction.service;
 
 import lombok.RequiredArgsConstructor;
+import org.spring.attraction.ENUM.AttractionMessage;
+import org.spring.attraction.ENUM.AttractionTypeMessage;
 import org.spring.attraction.dto.AttractionTypeDto;
+import org.spring.attraction.entity.Attraction;
 import org.spring.attraction.entity.AttractionType;
+import org.spring.attraction.entity.AttractionTypeList;
 import org.spring.attraction.repository.AttractionRepository;
+import org.spring.attraction.repository.AttractionTypeListRepository;
 import org.spring.attraction.repository.AttractionTypeRepository;
 import org.springframework.stereotype.Service;
 
@@ -15,11 +20,19 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class AttractionTypeService {
     private final AttractionTypeRepository attractionTypeRepository;
+    private final AttractionTypeListRepository attractionTypeListRepository;
 
-    public boolean save(AttractionTypeDto attractionTypeDto) {
-        AttractionType attractionType = AttractionType.toEntity(attractionTypeDto);
-        attractionTypeRepository.save(attractionType);
-        return true;
+    public AttractionTypeMessage save(AttractionTypeDto attractionTypeDto) {
+        AttractionTypeMessage result =  AttractionTypeDto.validate(attractionTypeDto);
+        if(result != null) {
+            return result;
+        }
+        AttractionType attractionType = attractionTypeRepository.findByType(attractionTypeDto.getType());
+        if(attractionType == null) {
+            attractionTypeRepository.save(AttractionType.toEntity(attractionTypeDto));
+            return AttractionTypeMessage.getTypeById(1);
+        }
+        return AttractionTypeMessage.getTypeById(-1);
     }
 
     public List<AttractionTypeDto> findAll() {
@@ -35,7 +48,13 @@ public class AttractionTypeService {
         return null;
     }
 
-    public void delete(Long id) {
+    public AttractionTypeMessage delete(Long id) {
+        List<AttractionTypeList> attractionTypeListList = attractionTypeListRepository.findByAttractionTypeId(id);
+        if(!attractionTypeListList.isEmpty()) {
+            return AttractionTypeMessage.getTypeById(-2);
+        }
         attractionTypeRepository.deleteById(id);
+
+        return AttractionTypeMessage.getTypeById(2);
     }
 }
